# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Arrivo API Documentation",
        "description": "API endpoints for Savings, Groups, and User Management",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "localhost",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/api/groups": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Get all groups the user belongs to",
                "operationId": "c19a6a9abc8c5a6d38d28caa82e8c646",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Create a new group",
                "operationId": "96d4e7ce5ba01d7408abc5d63a14e05c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "December Vacation"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/groups/{group}": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Get specific group details",
                "operationId": "505901cb360e8163f8aa3b7749a88321",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Groups"
                ],
                "summary": "Edit a group's information",
                "operationId": "2a4b639c66f230328a4ba814fd33f482",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/groups/{group}/friends/{friend}": {
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Add a friend to a specific group",
                "operationId": "984a3108d7be2a04d1dc7a70d8799050",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "friend",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Friend added"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Groups"
                ],
                "summary": "Remove a friend from a specific group",
                "operationId": "f4f66fa6725f92c6164ba047cfe905bd",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "friend",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend removed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/friends": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "List user friends",
                "operationId": "3fea10f4a8c6061e90b282f2d0d45866",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Add a friend",
                "operationId": "2e28cbda719b94a3bdb73bea54f0728f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "friend_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Friendship created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/friends/{friend}": {
            "delete": {
                "tags": [
                    "Friends"
                ],
                "summary": "Remove a friend",
                "operationId": "f2d1667d03998b2d3437553d1edb14cd",
                "parameters": [
                    {
                        "name": "friend",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend removed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/invites/{recipient}/invite": {
            "post": {
                "tags": [
                    "Invites"
                ],
                "summary": "Send group invite to a user",
                "operationId": "a9a58ba1d42471934c191d4b79f8caf6",
                "parameters": [
                    {
                        "name": "recipient",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "group_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Invite sent"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/invites/{invitation}/reply": {
            "post": {
                "tags": [
                    "Invites"
                ],
                "summary": "Reply to a group invitation",
                "operationId": "34afa46b660f3a415f78561c90fdef9b",
                "parameters": [
                    {
                        "name": "invitation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reply": {
                                        "type": "string",
                                        "enum": [
                                            "yes",
                                            "no"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reply processed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}